% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{compSCORE}
\alias{compSCORE}
\title{computes a dataframe of significant fitness scores from a matrix of significance scores
defined by a fitness score cutoff}
\usage{
compSCORE(mat, coln, sig = 1)
}
\arguments{
\item{mat}{matrix of fitness scores}

\item{coln}{column # of matrix with the sample of interest}

\item{sig}{= significance threshold, default = 1}
}
\value{
dataframe with 3 columns: gene, index, score. index indicating significant if 1,
nonsignificant if 0
}
\description{
computes a dataframe of significant fitness scores from a matrix of significance scores
defined by a fitness score cutoff
}
