% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dfGOBP}
\alias{dfGOBP}
\title{Dataframe listing Saccharomyces cerevisiae GOIDs and terms,
geneSetSizes, geneSet members}
\format{
Dataframe listing GOIDs and terms
     \describe{
         \item{GOID}{GO ID}
         \item{term}{GO term}
         \item{gene}{GO genesets}
         \item{geneSetSize}{GO geneset size}
}
}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\description{
Dataframe GOID terms correspond to names(yGOBP)
annotations were downloaded in January 2022 from Gene Ontolog
Consortium (link below)
}
