\name{goEnrichment-package}
\alias{goEnrichment-package}
\alias{goEnrichment}
\docType{package}
\title{
Enables easy and quick Gene Ontology (GO) enrichment analyses wrapping the
functions of GOstats and GSEABase
~~ goEnrichment ~~
}
\description{
Enables easy and quick Gene Ontology (GO) enrichment analyses wrapping the
functions of GOstats and GSEABase
}
\details{
\tabular{ll}{
Package: \tab goEnrichment\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-04\cr
License: \tab GPL-3\cr
}
~~ Usage ~~
Find out the path to the installed goEnrichment package: \code{path.package("goEnrichment")}
Copy the example input file and edit it to your needs \code{path/2/package/input.R}
}
\author{
Maintainer: Dr. Asis Hallab <hallab@mpipz.mpg.de>
}
