% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrichment.R
\docType{package}
\name{GOenrichment}
\alias{GOenrichment-package}
\alias{GOenrichment}
\title{GOenrichment: A Package for GO Enrichment Visualization}
\description{
Provides functions for visualizing Gene Ontology (GO) enrichment results using
`visNetwork` and `igraph`. It includes tools for graph creation, layout optimization,
and interactive visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gurinina/GOenrichment}
  \item Report bugs at \url{https://github.com/gurinina/GOenrichment/issues}
}

}
