% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/goEnrichment.R
\name{goEnrichTest}
\alias{goEnrichTest}
\title{Computes a Gene Ontology (GO) enrichment test using the functions from
package GOstats and GSEABase.}
\usage{
goEnrichTest(gsc, gene.ids, univ.gene.ids, ontologies = c("BP", "CC", "MF"),
  pvalue.cutoff = 0.01, cond = FALSE, test.dir = "over",
  p.adjust.method = "fdr")
}
\arguments{
\item{gsc}{An instance of gene set collection [package GSEABase]}

\item{gene.ids}{A character vector of gene IDs among which to look for
enriched GO terms}

\item{univ.gene.ids}{A character vector of 'background' gene IDs to be used
as reference GO annotations when looking for overrepresentations in
'gene.ids'}

\item{ontologies}{A character vector of the Gene Ontology categories to
compute enrichment tests for, default is 'BP', 'CC', and 'MF'}

\item{pvalue.cutoff}{The significance level, default to 0.01}

\item{cond}{If set to TRUE already found to be enriched GO annotations wont
be counted when testing enrichment for parental terms, default is
FALSE}

\item{test.dir}{Set to 'over' to infer overrepresentation, set to 'under'
if the opposite is wanted.}

\item{p.adjust.method}{The method used to adjust p-values for multiple
       hypothesis testing, default is 'fdr'. See function p.adjust for
       details.}
}
\value{
A named list with enriched GO terms in the argument GO categories.
}
\description{
Computes a Gene Ontology (GO) enrichment test using the functions from
package GOstats and GSEABase.
}

