% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
\name{hGOBP.gmt}
\alias{hGOBP.gmt}
\title{BP genesets from Homo sapiens}
\format{
gmt file: a list of BP genesets where names = GO term
}
\source{
\url{http://baderlab.org/GeneSets}
}
\description{
GO biological process (BP) genesets from
Downloaded from Gary Bader's lab in July 2022
}
