% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{hyperG}
\alias{hyperG}
\title{computes enrichment using the hypergeometric test given:}
\usage{
hyperG(
  querySet,
  geneSets,
  uni,
  scoreMat,
  minSetSize = minGeneSetSize,
  maxSetSize = maxGeneSetSize,
  uniSize = NA
)
}
\arguments{
\item{querySet}{character vector of genes in query set, genes with significant fitness defect scores}

\item{geneSets}{named list of gene sets to test for significant overlap w/ the query set}

\item{uni}{character vector of genes in the universe (i.e. background set)
- if NULL, must specify uniSize}

\item{scoreMat}{dataframe of gene scores
- first column = gene column
- second column = index, 1 and 0s designating significance
- third column = fitness scores, descending order
- can be NULL}

\item{maxSetSize}{max number of genes in geneSets (after restricting to the gene universe)}

\item{uniSize}{the number of genes in the universe}
}
\value{
dataframe of enrichment results, sorted by increasing FDR value. The columns are:
        querySetFraction = the fraction of the query set that overlaps with the term set
        geneSetFraction = the fraction of the term set that overlaps with the query set
        foldEnrichment = the fold enrichment of the query set with the term genes
        P = P value estimating the significance with which the query set is enriched with the term genes
        FDR = FDR value estimating the significance of enrichment
        overlapGenes = a |-separated list of genes in the overlap of the query set and the term set;
                       if scoreMat is provided (not NULL), the scores of the genes are shown in parentheses
       maxOverlapGeneScore = if scoreMat is provided (not NULL), the maximum score of the overlapGenes
}
\description{
computes enrichment using the hypergeometric test given:
}
