% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/goEnrichment.R
\name{procGOHyperGResult}
\alias{procGOHyperGResult}
\title{Converts an instance of GOHyperGResult into a table containing only the
significantly enriched Gene Ontology terms.}
\usage{
procGOHyperGResult(ghgr, pvalue.cutoff = 0.01)
}
\arguments{
\item{ghgr}{The instance of GOHyperGResult to process}

\item{pvalue.cutoff}{The significance level to be applied}
}
\value{
A data.frame with three columns: 1. 'GO.term', 2. 'p.value', and 3.
        'GO.category' - or NULL if no significantly enriched GO terms were
        found.
}
\description{
Converts an instance of GOHyperGResult into a table containing only the
significantly enriched Gene Ontology terms.
}

