% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{runNetwork}
\alias{runNetwork}
\title{runNetwork}
\usage{
runNetwork(
  nodes,
  edges,
  height = 1300,
<<<<<<< HEAD
  main = list(text = nodes$filename[1], style = "font-size:40px;text-align:center;")
=======
  main = list(text = nodes$filename[1], style = "font-size:40px;text-align:center;"),
  ...
>>>>>>> e3df6ff77d1d96c2a2944bb6d9104fa2f8d15259
)
}
\arguments{
\item{nodes}{dataframe returned from visSetup, "nodes" see ?? visSetup}

\item{edges}{dataframe returned from visSetup, "edges" see ?? visSetup}

\item{height}{A numeric value}
<<<<<<< HEAD

\item{main}{character title list, includes font size and color if desired}
=======
>>>>>>> e3df6ff77d1d96c2a2944bb6d9104fa2f8d15259
}
\value{
network
}
\description{
Uses the output from visSetup to generate a network of gene enrrichments.
}
\details{
The larger a node in the netwark, the more significance the
GO enrichment for that term. The thicker the edge connecting twa nodes,
the greater the overlap betweem twp terms. Bold and enlarged font size
indicate the most enriched term for a given cluster.
}
