% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
\name{sampleFITdata}
\alias{sampleFITdata}
\title{Fitness data from Saccharomyces cerevisiae}
\format{
numeric matrix with two columns; 4956 gene names in rows
}
\source{
\url{https://www.ggcnlab.com/}
}
\description{
Experimental data from G. Giaever & C. Nislow labs
}
