% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{visSetup}
\alias{visSetup}
\title{visSetup}
\usage{
visSetup(enrichInfo, edgeMat, fontsize = 22, fontface = "Arial")
}
\arguments{
\item{enrichInfo}{a data frame, see ??runGORESP enrichInfo output for explanation of these columns}

\item{edgeMat}{a data frame, see ??runGORESP edgeMat output for explanation of these columns}

\item{fontsize}{fontsize}

\item{fontface}{fontface}
}
\value{
returns two dataframes in a list, vis$nodes with the following columns:
id, filename,term, nGenes, nQuery, nOverlap, querySetFraction, geneSetFraction, foldEnrichment,
P, FDR, overlapGenes, maxOverlapGeneScore: see ??runGORESP enrichInfo output for explanation of these columns
visualization columns: size, formattedLabel, label, color.background,color.border,
borderWidthSelected, color.highlight.border, color.highlight.background,
color.hover.background, color.hover.border, font.face, shape, font.size, font.bold,
borderWidth, labelHighlightBold and and vis$edges with the columns from, to, overlapCoeff,
width and color
}
\description{
this function takes the output from runGORESP generates objects for drawing GO enrichment networks
}
