% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
\name{yGOBP.gmt}
\alias{yGOBP.gmt}
\title{Saccharomyces GO biological process (BP) genesets}
\format{
gmt file: a list of BP genesets where names = GO term
}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\description{
BP genesets were downloaded in January 2022
from the Gene Ontolog Consortium (link below).
}
