% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{compSCORE}
\alias{compSCORE}
\title{compSCORE:}
\usage{
compSCORE(mat, coln, sig = 1)
}
\arguments{
\item{mat}{numeric matrix of fitness scores}

\item{coln}{numeric, column # of matrix with the sample of interest}

\item{sig}{numeric, significance threshold, default = 1}
}
\value{
dataframe with 3 columns: gene, index, score. index indicating significant 1,
or nonsignificant, 0
}
\description{
computes a dataframe of designating significant fitness scores from a
matrix of screening data
}
