% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
\name{dfGOBP}
\alias{dfGOBP}
\title{BP genesets from Saccharomyces cerevisiae}
\format{
Dataframe listing GOIDs and terms
  \describe{
     \item{GOID}{GO ID}
     \item{term}{GO term}
     \item{gene}{GO genesets}
     \item{geneSetSize}{GO geneset size}
}
}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\description{
GO biological process (BP) genesets from
Saccharomyces cerevisiae downloaded in January 2022
}
