% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrichment.R
\docType{package}
\name{GOenrichment-package}
\alias{GOenrichment}
\alias{GOenrichment-package}
\title{GOenrichment: A Package for GO Enrichment Visualization}
\description{
Provides functions for visualizing GO enrichment results using `visNetwork` and `igraph`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gurinina/GOenrichment}
  \item Report bugs at \url{https://github.com/gurinina/GOenrichment/issues}
}

}
