% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
\name{hGOBP.gmt}
\alias{hGOBP.gmt}
\title{Homo sapiens GO biological process (BP) genesets}
\format{
gmt file: a list of BP genesets where names = GO term
}
\source{
\url{http://baderlab.org/GeneSets}
}
\description{
BP genesets were downloaded from Gary Bader's
lab in July 2022: human_go_bp_no_go_iea_entrezgene.gmt
(biological process minus those annoted using evidence code
inferred by electronic annotation (IEA))
}
