% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/goEnrichment.R
\name{joinGOEnrichResults}
\alias{joinGOEnrichResults}
\title{Method extracts all GO ontology specific enrichment result and combines them
into a single data frame, re-adjusting the Pvalues for multiple hypothesis
testing.}
\usage{
joinGOEnrichResults(go.enrich.lst, ontology.tbl.entry = "hyperGTest.table",
  p.adjust.method = "fdr")
}
\arguments{
\item{go.enrich.lst}{The result of calling method goEnrichTest(…).}

\item{ontology.tbl.entry}{The name of the list-entry in which to find the
go-enrichment result as table.}

\item{p.adjust.method}{The method to use when adjusting p-values for
       multiple hypothesis testing.}
}
\value{
A data.frame with the merged enriched GO terms and adjusted
p-values.
}
\description{
Method extracts all GO ontology specific enrichment result and combines them
into a single data frame, re-adjusting the Pvalues for multiple hypothesis
testing.
}

