% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{runGORESP}
\alias{runGORESP}
\title{@param mat matrix of fitness data
    @param coln the column of the matrix with sample of interest
    @param curr_exp name of exp, usually after the sample name so set to colnames(mat)[coln]
    @param sig significance threshold of fitness defect score, default is 1
    @param fdrThresh fdr threshold for significance cutoff of enrichments, default = 0.2
    @param bp_path path of gmt file
    @param bp_input alternatively th gmt file itself
    @param go_path path of file with GOID and terms path of gmt file alternatively bp_input, gmt file itself,
    @param go_input alternatively the file with GOID and terms itself
    @param minGeneSetSize lower limit on geneSetSize, default 5
    @param maxGeneSetSize upper limit on geneSetSize,,default 300}
\usage{
runGORESP(
  mat,
  coln,
  curr_exp = colnames(mat)[coln],
  sig = 1,
  fdrThresh = 0.2,
  bp_path = NULL,
  bp_input = GOBPgmt,
  go_path = NULL,
  go_input = dfGOBP,
  minGeneSetSize = 5,
  maxGeneSetSize = 300
)
}
\description{
RETURNS two dataframes of enrichment results, sorted by increasing FDR value.
}
\details{
@return enrichInfo dataframe includes the columns: filename,term, nGenes, nQuery, nOverlap,
    querySetFraction, geneSetFraction, foldEnrichment, P, FDR, overlapGenes, maxOverlapGeneScore and columns used for visuealization purposes such as
    cluster, id, size and formattedLabel

    querySetFraction: the fraction of the query set (the set that passes the significance threshold)
     that overlaps with the term set
    geneSetFraction: the fraction of the term set that overlaps with the query set
    foldEnrichment = the fold enrichment of the query set over the bgRate, where bgRate
       is nGenes in geneSet/genes in the universe
    P = P value estimating the significance with which the query set is enriched with the term genes
    FDR = FDR value estimating the significance of enrichment
    overlapGenes = a |-separated list of genes that overlap the query set and the geneSet
                       if scoreMat is provided (not NULL), the scores of the genes are shown in parentheses
   maxOverlapGeneScore = if scoreMat is provided (not NULL), the maximum score of the overlapGenes
    @return edgeMat dataframe includes the columns:
            source, target, overlapCoeff, width and label. These map the terms from the enrichInfo
            dataframe and serve to show the overlap between terms for downstream visualization purposes
            and for understanding the relationship between terms;
}
