% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGORESP.R
\name{runGORESP}
\alias{runGORESP}
\title{GO enrichment analysis using the hypergeometric test}
\usage{
runGORESP(
  scoreMat,
  curr_exp = "test",
  fdrThresh = 0.2,
  bp_path = NULL,
  bp_input = NULL,
  go_path = NULL,
  go_input = NULL,
  minSetSize = 5,
  maxSetSize = 300
)
}
\arguments{
\item{scoreMat}{Data frame with gene scores. The first column is "gene," the second is "index" (indicating significance with 1 or 0), and the third is "score" (fitness scores in descending order).}

\item{curr_exp}{Character; experiment label for the analysis, default = "test".library(GOenrichment)}

\item{fdrThresh}{Numeric; FDR threshold for enrichment significance, default = 0.2.}

\item{bp_path}{Character; path to a .gmt file for biological process (BP) gene sets.}

\item{bp_input}{List; gene sets in .gmt format. If provided, `bp_path` is ignored.}

\item{go_path}{Character; path to file with GO terms and IDs, or NULL if using `go_input`.}

\item{go_input}{Data frame; GO terms and IDs. If provided, `go_path` is ignored.}

\item{minSetSize}{Numeric; minimum number of genes in a gene set for inclusion, default = 5.}

\item{maxSetSize}{Numeric; maximum number of genes in a gene set for inclusion, default = 300.}

\item{uniSize}{Numeric; number of genes in the universe, or length of `uni`.}
}
\value{
List containing:
\describe{
  \item{enrichInfo}{Data frame with GO enrichment results including columns:}
  \describe{
    \item{term}{Name of the GO term.}
    \item{querySetFraction}{Fraction of the query set overlapping with the term set.}
    \item{geneSetFraction}{Fraction of the term set overlapping with the query set.}
    \item{foldEnrichment}{Fold enrichment of query genes within the GO term genes.}
    \item{P}{P-value for enrichment significance.}
    \item{FDR}{False discovery rate (adjusted P-value) for enrichment significance.}
    \item{overlapGenes}{List of overlapping genes in the query and term sets.}
    \item{maxOverlapGeneScore}{Maximum fitness score among overlapping genes, if `scoreMat` is provided.}
    \item{cluster, id, size, formattedLabel}{Additional clustering and visualization information.}
  }
  \item{edgeMat}{Data frame for gene set term overlaps, including source, target, and overlap coefficients for visualization.}
}
}
\description{
Performs a GO enrichment analysis using the hypergeometric test for a set of query genes that pass a user-defined fitness score threshold. It compares the query genes against a background set (universe) and outputs enrichment and clustering information.
}
\examples{
# Example usage:
scoreMat <- compSCORE(mat, coln, sig = sig)
curr_exp <- colnames(mat)[coln]
hresp <- GOenrichment::runGORESP(fdrThresh = 0.2, scoreMat = scoreMat,
  coln = 1, curr_exp = "test", sig = 1,
  bp_input = hGOBP.gmt, go_input = NULL, minSetSize = 20,
  maxSetSize = 300)
}
