% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_March13_GORESP.R
\name{runNetwork}
\alias{runNetwork}
\title{runNetwork: takes the output from visSetup generating a network of nodes comprised of
geheSets defined by Gene Ontology (GO) biological processes. Analysis is restricted
to genesets comprised at least 5 (minGeneSetSize) and not more than 300 genes (maxGeneSetSize).}
\usage{
runNetwork(nodes, edges)
}
\value{
network
The significance of the enrichments was estimated using the hypergeometric test
for the set of genes passing the fitness score threshold.connected by edges.
The enrichment of a GO gene set is driven by the gene subset also present in the
query set derived from the genes passing the fitness score threshold. In some cases,
a similar set of query set genes drives the enrichment of multiple GO gene sets.
Of these GO sets, we highlight the most significantly enriched, and the rest are
considered redundant.
}
\description{
runNetwork: takes the output from visSetup generating a network of nodes comprised of
geheSets defined by Gene Ontology (GO) biological processes. Analysis is restricted
to genesets comprised at least 5 (minGeneSetSize) and not more than 300 genes (maxGeneSetSize).
}
