% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNetwork.R
\name{runNetwork}
\alias{runNetwork}
\title{Visualize a gene network using visNetwork
This function visualizes a gene network from GO enrichment analysis results.
It requires pre-processed node and edge data to create an interactive network plot.}
\usage{
runNetwork(
  nodes,
  edges,
  height = 1300,
  main = "Network of GO Enrichment Results"
)
}
\arguments{
\item{nodes}{data.frame; contains node information for the network (see `runGORESP` enrichInfo output).}

\item{edges}{data.frame; contains edge information for the network (see `runGORESP` edgeMat output).}

\item{height}{Numeric; height of the network plot in pixels. Default is 1300.}

\item{main}{List; main title settings with `text` and `style`. Default is `list(text = nodes$filename[1], style = "font-size:40px;text-align:center;")`}
}
\value{
No return value, called for side effects. Generates an interactive network visualization in the viewer.
}
\description{
Visualize a gene network using visNetwork
This function visualizes a gene network from GO enrichment analysis results.
It requires pre-processed node and edge data to create an interactive network plot.
}
\examples{
# Assuming `nodes` and `edges` are data frames obtained from `runGORESP`
runNetwork(nodes, edges, height = 1300, main = list(text = nodes$filename[1], style = "font-size:40px;text-align:center;"))
}
