% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNetwork.R
\name{runNetwork}
\alias{runNetwork}
\title{Generates a network of GO enrichments.using the output from visSetup}
\usage{
runNetwork(
  nodes,
  edges,
  height = 1300,
  main = list(text = nodes$filename[1], style = "font-size:40px;text-align:center;")
)
}
\arguments{
\item{nodes}{dataframe returned from visSetup, "nodes" see ?? visSetup}

\item{edges}{dataframe returned from visSetup, "edges" see ?? visSetup}

\item{height}{A numeric value}

\item{main}{character title list, includes font size and color if desired}
}
\value{

}
\description{
Generates a network of GO enrichments.using the output from visSetup
}
\examples{
# Example usage:
scoreMat = compSCORE(mat, coln, sig = sig)
curr_exp = colnames(mat)[coln]
hresp <- GOenrichment::runGORESP(fdrThresh = 0.2, scoreMat = scoreMat,
  coln=1, curr_exp = curr_exp, sig = 1,
  bp_input = hGOBP.gmt, go_input = NULL, minSetSize = 20,
  maxSetSize = 300)
}
