% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
<<<<<<< HEAD
\name{sampleFITdata}
\alias{sampleFITdata}
\title{sample fitness data from Saccharomyces cerevisiae}
\format{
numeric 4956 x 2 matrix, genes in rows,sample in columns
=======
<<<<<<< HEAD
\name{sampleFitdata}
\alias{sampleFitdata}
\title{Fitness data from Saccharomyces cerevisiae}
\format{
numeric matrix with two columns; 4956 gene names in rows
        num [1:4956, 1:2] -0.0806 0.038 NA 0.7913 0.0984 ...
      \describe{
     \item{chr [1:4956]}{"AAC1" "AAC3" "AAD14" "AAD3" ...}
     \item{chr [1:2]}{"azithromycin_1.75mM" "azithromycin_2mM"}
}
=======
\name{sampleFITdata}
\alias{sampleFITdata}
\title{Fitness data from Saccharomyces cerevisiae}
\format{
numeric matrix with two columns; 4956 gene names in rows
>>>>>>> e3df6ff77d1d96c2a2944bb6d9104fa2f8d15259
>>>>>>> fecbf8780150295844477dc4753b3f7b7aa8037c
}
\source{
\url{https://www.ggcnlab.com/}
}
\description{
HOP chemogenomic screening data from G. Giaever & C. Nislow labs
}
\details{
Chemogenomic Profilimg data from Saccharomyces cerevisiae
}
\author{
G.Giaever & C.Nislow
}
\keyword{chemogenomics,}
\keyword{competitive}
\keyword{fitness}
\keyword{profiling}
