% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleFITdata}
\alias{sampleFITdata}
\title{sample fitness data from Saccharomyces cerevisiae}
\format{
numeric 4956 x 2 matrix, genes in rows,sample in columns
}
\source{
\url{https://www.ggcnlab.com/}
}
\description{
HOP chemogenomic screening data from G. Giaever & C. Nislow labs
}
\details{
Chemogenomic Profilimg data from Saccharomyces cerevisiae
}
\author{
G.Giaever & C.Nislow
}
\keyword{chemogenomics,}
\keyword{competitive}
\keyword{fitness}
\keyword{profiling}
