% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleFITdata}
\alias{sampleFITdata}
\title{sample fitness data from Saccharomyces cerevisiae}
\source{
\url{https://www.ggcnlab.com/}
}
\description{
HOP chemogenomic screening data from G. Giaever & C. Nislow labs
}
