% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetup.R
\name{visSetup}
\alias{visSetup}
\title{Prepare network data for gene visualization using visNetwork}
\usage{
visSetup(enrichInfo, edgeMat, fontsize = 22, fontface = "Arial")
}
\arguments{
\item{enrichInfo}{Data frame of enrichment information, as produced by `runGORESP`.}

\item{edgeMat}{Data frame of edges, showing relationships between nodes, as produced by `runGORESP`.}

\item{fontsize}{Numeric; font size for node labels. Default is 22.}

\item{fontface}{Character; font face for node labels. Default is "Arial".}
}
\value{
A list with two formatted data frames: `nodes` (containing attributes for each node) and `edges` (representing relationships between nodes), both suitable for `visNetwork` plotting.
}
\description{
Prepare network data for gene visualization using visNetwork
}
\examples{
visSetup(enrichInfo, edgeMat)
}
