% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visSetup.R
\name{visSetup}
\alias{visSetup}
\title{Calculate Significant Fitness Scores
this function
The `visSetup` function takes the output from runGORESP and generates objects for drawing GO enrichment networks}
\usage{
visSetup(enrichInfo, edgeMat, fontsize = 22, fontface = "Arial")
}
\arguments{
\item{enrichInfo}{A data frame, see ?runGORESP enrichInfo output for explanation of these columns}

\item{edgeMat}{A data frame, see ?runGORESP edgeMat output for explanation of these columns}

\item{fontsize}{fontsize for node labels; default = 22}

\item{fontface}{fontface for node labels; default = "Arial"}
}
\value{
Returns two dataframes in a named list for using with visNetwork; Network visualization using vis.js library.
nodes has the following columns: id, filename,term, nGenes, nQuery, nOverlap, querySetFraction, geneSetFraction, foldEnrichment, P, FDR, overlapGenes, maxOverlapGeneScore: see ??runGORESP enrichInfo output for explanation of these columns
\describe{
  \item{nodes}{A binary indicator of significance: 1 if the score is significant, 0 otherwise.}
  \item{score}{The fitness score of the gene in the specified column.}
  \item{gene}{The name of the gene, taken from the row names of `mat`.}
}
}
\description{
Calculate Significant Fitness Scores
this function
The `visSetup` function takes the output from runGORESP and generates objects for drawing GO enrichment networks
}
\examples{
# Example usage:
scoreMat = compSCORE(mat, coln, sig = sig)
curr_exp = colnames(mat)[coln]
hresp <- GOenrichment::runGORESP(fdrThresh = 0.2, scoreMat = scoreMat,
  coln=1, curr_exp = curr_exp, sig = 1,
  bp_input = hGOBP.gmt, go_input = NULL, minSetSize = 20,
  maxSetSize = 300)
vis <-  visSetup(enrichInfo = hresp$enrichInfo,edgeMat = hresp$edgeMat)
}
