% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.R
\name{yGOBP.gmt}
\alias{yGOBP.gmt}
\title{BP genesets from Saccharomyces cerevisiae}
\format{
gmt file: a list of BP genesets where names = GO term
}
\source{
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\description{
GO biological process (BP) genesets from
Saccharomyces cerevisiae downloaded in January 2022
}
